Add test for shoot method
