+++ src/Board_test.py
@@ -2,6 +2,8 @@ import unittest
 import os
 
 from Board import initBoard, getBoardFromConfig
+from constants import PLAYER_1, PLAYER_2, FLEET_LIFE
+import ui
 
 TESTDATA_FILENAME = os.path.join(os.path.dirname(__file__), '../config/player1')
 
@@ -19,9 +21,6 @@ class TestMainMethods(unittest.TestCase):
         self.assertEqual(initBoard(5), board)
 
     def test_board_config(self):
-        
-        # os.path.relpath('..\\subfldr1\\testfile.txt', os.path.dirname(__file__))
-        # f = open("./config/player1", "r")
         with open("./config/player1", 'r') as f:
             config = f.read().splitlines()
